requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
python-telegram-bot==21.5
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
